# Make usec parsing faster